Move oracle sqlite handling into utils/sqlite.py
